requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai>=1.6.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
Pillow==10.0.0
python-decouple==3.8
six==1.16.0
